pyproject.toml
[project]
name = "photo-restore-batch"
version = "0.1.0"
description = "Batch photo restoration pipeline (normalization stage)"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26.0",
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
face-enhance = [
  "gfpgan==1.3.8",
  "basicsr==1.4.2",
  "facexlib==0.3.0",
  "torch==2.10.0",
  "torchvision==0.25.0",
  "opencv-python==4.13.0.92",
]

[project.scripts]
restore-batch = "restore_batch.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
